fix: broken ids, missing awaits
